requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.59.0
h11==0.9.0
httptools==0.1.1
idna==2.10
pydantic==1.6
PyYAML==5.3.1
requests==2.24.0
starlette==0.13.4
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
